Reorganize Flux API utilities
