+++ ui/src/shared/apis/flux/suggestions.ts
@@ -1,0 +1,14 @@
+import AJAX from 'src/utils/ajax'
+
+export const getSuggestions = async (url: string) => {
+  try {
+    const {data} = await AJAX({
+      url,
+    })
+
+    return data.funcs
+  } catch (error) {
+    console.error('Could not get suggestions', error)
+    throw error
+  }
+}
